size

+++ 0x00-hello_world/6-size.c
@@ -3,7 +3,7 @@
 /**
  * main - prints the sieze of many var types
  *
- * Return: always 0 (success)
+ * Return: Always 0 (success)
  */
 int main(void)
 {
@@ -13,11 +13,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("size of char: %lu byte(s)\n", sizeof(a));
-	printf("size of an int: %lu byte(s)\n", sizeof(b));
-	printf("size of a long int: %lu byte(s)\n", sizeof(c));
-	printf("size of long long int: %lu byte(s)\n", sizeof(d));
-	printf("size of a float: %lu bytes(s)\n", sizeof(e));
+	printf("size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("size of a float: %lu bytes(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
-
